filetests: allow for running clift filetests on non-host ISA
zkasm supports running tests on any architecture that can run nodejs
which is somewhat unique in terms of ISAs that cranelift filetest runner
supports.

For now things still do not work though, as there appear to be attempts
to generate a trampoline. That’s becasue the rest of the infrastructure
still assumes this is going to be JITd, loaded into memory directly and
called… directly. Whereas for zkasm we must call an external process,
unfortunately.