Fix memory operations with external call
This is not a full fix, as it relies on the unverified external call, but it gets the job done for making memory operations execute correctly.
The challenge right now is that addresses are 32-bit integers and we need to shift them back before doing the memory access.
I plan to try migrating to the masking-based approach of handling 32 bit integers, which would remove the need to do the external call.