Introduce ZkasmEnvironment
This PR forks DummyEnvironment that we used before, as we will need to change it to enable memory operations.
At this point, there are no changes to the environment and this should be a no-op.

In the past, we relied on DummyEnvironment which is only useful in testing, but in reality, we need a dedicated environment suited for ZKASM machine.