Use constant base pointers instead of vmctx
In ZKASM machine we have all memory dedicated to a single program, so there is no need for a separate base pointer and we can use constant offsets instead.

I haven't found a way to pass a literal constant to HeapData and needed to create an intermediate ir::GlobalValueData::Symbol that gets resolved to a constant during lowering.
Ideally, we would find a way to use a constant as that would make codegen fold that constant during accesses.