Implement global data segments
I haven't found a good way to generate a custom AMode for these accesses
to use global ZKASM arrays, as we don't get to decide how these data
segment acceses are generated - this is done in Rustc and we have a very
limited control here. But at least we have a guarantee that Rust will
not generate conflicting data acceses here, which should be good enough
for now.

In SHA256 example specifically, data offset starts from 1MB because the
memory before that is allocated for stack. This can be controlled by the
Rustc compilation flag and we need to think a bit more about the value
that we want to use there.