Align stack handling with other architectures
In the past the stack in ZKASM grew from lower to upper addresses, but
there is no real reason for this, as we can always start from the
largest offset and go down.

This aligns the behaviour with other architectures which makes some
common logic like spill stack allocation work correctly.

We also now track all offsets in bytes and only do conversion to slots on
accesses to make sure that beforementioned spill slock logic works
correctly.