Add Rust zkAsm benchmarks
I've tried to polish the approach based on build-artifacts but it was too inflexible - I wasn't able to fully customize build profile and flags used by the artifact. On top of that, it required using nightly Cargo which generated tons of warnings.

So I migrated the implementation to build.rs based on the approach taken by https://github.com/near/nearcore/tree/master/runtime/near-test-contracts. This worked like a charm and I think it's ready for a high-level review.

Based on artifact-dependencies to compile Rust to WASM during build.